{"schema":"decodex/commit/1","summary":"Modularize host FFI session bridge","authority":"manual"}
